+++ src/lib/proof_system.ts
@@ -18,6 +18,8 @@ import {
   FlexibleProvablePure,
   InferProvable,
   ProvablePureExtended,
+  Struct,
+  provable,
   provablePure,
   toConstant,
 } from './circuit_value.js';
@@ -25,7 +27,7 @@ import { Provable } from './provable.js';
 import { assert, prettifyStacktracePromise } from './errors.js';
 import { snarkContext } from './provable-context.js';
 import { hashConstant } from './hash.js';
-import { MlArray, MlBool, MlResult, MlPair, MlUnit } from './ml/base.js';
+import { MlArray, MlBool, MlResult, MlPair } from './ml/base.js';
 import { MlFieldArray, MlFieldConstArray } from './ml/fields.js';
 import { FieldConst, FieldVar } from './field.js';
 import { Cache, readCache, writeCache } from './proof-system/cache.js';
@@ -47,6 +49,7 @@ export {
   Empty,
   Undefined,
   Void,
+  VerificationKey,
 };
 
 // internal API
@@ -261,7 +264,7 @@ function ZkProgram<
 ): {
   name: string;
   compile: (options?: { cache?: Cache; forceRecompile?: boolean }) => Promise<{
-    verificationKey: { data: string; hash: Field } 
+    verificationKey: { data: string; hash: Field };
   }>;
   verify: (
     proof: Proof<
@@ -483,6 +486,13 @@ class SelfProof<PublicInput, PublicOutput> extends Proof<
   PublicInput,
   PublicOutput
 > {}
+
+class VerificationKey extends Struct({
+  ...provable({ data: String, hash: Field }),
+  toJSON({ data }: { data: string }) {
+    return data;
+  },
+}) {}
 
 function sortMethodArguments(
   programName: string,